Renamed inbox-panel and the cards config file
