+++ client/header/activity-panel/panels/inbox/inbox.scss
@@ -4,7 +4,7 @@
 		margin-top: $gap-large;
 	}
 
-	.woocommerce-abbreviated-panels {
+	.woocommerce-abbreviated-cards {
 		border-top: 1px solid $gray-200;
 		& > div {
 			border: none;
